feat: Rename process to proc, which sounds less like an OS process
